requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
diff-match-patch==20241021
Django==5.2.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-datatables-view==1.20.0
django-extensions==4.1
django-import-export==4.3.7
django-select2==8.4.0
django-tables2==2.7.5
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
